Move GPU solver under classical folder [by openclaw: vitamin]
