+++ DMET/ProblemSolver/__init__.py
@@ -3,7 +3,7 @@ from .ClassicalEigenSolver.EigenSolver import EigenSolver
 from .QuantumEigenSolver.EigenSolver import EigenSolver
 # GPU solver (optional)
 try:
-    from .GpuEigenSolver import EigenSolver as GpuEigenSolver
+    from .ClassicalEigenSolver.GpuEigenSolver import EigenSolver as GpuEigenSolver
 except Exception:
     # If GPU solver cannot be imported (no cupy), ignore
     GpuEigenSolver = None
